Split code in 2 seperate files
util_linux.go for all linux util functions, without cgo specific code
util_linux_cgo.go for all linux util functions depending on cgo

Reason:

github.com/sl1pm4t/terraform-provider-lxd uses the shared lxd code for
interactions with the lxd daemon. Compiling this for other
architectures, results in a build where archive_linux.go isn't
compilable because of missing functions defined in util_linux.go

This would fix this compile issue

```
gox
Number of parallel builds: 7

-->     windows/386: github.com/sl1pm4t/terraform-provider-lxd
-->       linux/arm: github.com/sl1pm4t/terraform-provider-lxd
-->   freebsd/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->   windows/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->      darwin/386: github.com/sl1pm4t/terraform-provider-lxd
-->     openbsd/386: github.com/sl1pm4t/terraform-provider-lxd
-->     freebsd/386: github.com/sl1pm4t/terraform-provider-lxd
-->       linux/386: github.com/sl1pm4t/terraform-provider-lxd
-->    darwin/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->     linux/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->      netbsd/arm: github.com/sl1pm4t/terraform-provider-lxd
-->     freebsd/arm: github.com/sl1pm4t/terraform-provider-lxd
-->      netbsd/386: github.com/sl1pm4t/terraform-provider-lxd
-->    netbsd/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->   openbsd/amd64: github.com/sl1pm4t/terraform-provider-lxd
-->     linux/s390x: github.com/sl1pm4t/terraform-provider-lxd
```
